split buildNetwork to 2 files
